# docs/swagger.yaml
basePath: /api/
definitions:
  types.AddUserRequestBody:
    properties:
      address:
        $ref: '#/definitions/types.Address'
      name:
        type: string
      phoneNumber:
        type: string
      username:
        type: string
    type: object
  types.Address:
    properties:
      city:
        type: string
      geo:
        $ref: '#/definitions/types.Geo'
      street:
        type: string
      suite:
        type: string
      zipcode:
        type: string
    type: object
  types.Geo:
    properties:
      lat:
        type: string
      lng:
        type: string
    type: object
  types.UpdateUserRequestBody:
    properties:
      address:
        $ref: '#/definitions/types.Address'
      name:
        type: string
      phoneNumber:
        type: string
      username:
        type: string
    type: object
  types.User:
    properties:
      address:
        $ref: '#/definitions/types.Address'
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: bbombelka@github.com
  description: This is a sample server for demonstrating Swagger in Go, showing how
    to add, get, update, and delete users from a contact book.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Users Contact Book API
  version: 1.0.0
paths:
  /users:
    get:
      description: Get details of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.User'
            type: array
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a user and returns the details
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.AddUserRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
      summary: Delete a user
      tags:
      - users
    get:
      description: get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a single user by id
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update a user's details by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.UpdateUserRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
      summary: Update a user
      tags:
      - users
schemes:
- http
- https
swagger: "2.0"