# main.go
package main

import (
	"encoding/json"
	"net/http"

	"github.com/bbombelka/contactbook/types"
	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"

	_ "github.com/bbombelka/contactbook/docs" // docs is generated by Swag CLI
)

//	@title			Users Contact Book API
//	@version		1.0.0
//	@description	This is a sample server for demonstrating Swagger in Go, showing how to add, get, update, and delete users from a contact book.
//	@termsOfService	http://example.com/terms/

//	@contact.email	bbombelka@github.com

//	@license.name	Apache 2.0
//	@license.url	http://www.apache.org/licenses/LICENSE-2.0.html

//	@host		localhost:8080
//	@BasePath	/api/

// @schemes	http https
func main() {
	r := gin.Default()

	initDatabase()

	// try adding base path later :)
	r.GET("/api/users", getAllUsersHandler)
	r.GET("/api/users/:id", getUserHandler)
	r.POST("/api/users", addUserHandler)
	r.DELETE("/api/users/:id", deleteUserHandler)
	r.PATCH("/api/users/:id", updateUserHandler)

	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	r.Run()
}

// GetUser godoc
// @Summary      Get a single user by id
// @Description  get user by ID
// @Tags         users
// @Produce      json
// @Param        id   path      int  true  "User ID"
// @Success      200  {object}   types.User
// @Failure      404  {object}   map[string]string
// @Router       /users/{id} [get]
func getUserHandler(c *gin.Context) {
	userId := extractUserId(c)

	user, hasFoundUser := findUser(userId)

	if hasFoundUser {
		c.JSON(200, user)
	} else {
		c.JSON(404, gin.H{"Message": "No user with this id found !"})
	}
}

// GetAllUsers godoc
// @Summary      Get all users
// @Description  Get details of all users
// @Tags         users
// @Produce      json
// @Success      200  {array}  types.User
// @Router       /users [get]
func getAllUsersHandler(c *gin.Context) {
	intLimit, intSkip := extractQueryParams(c)
	usersData := findUsers(intLimit, intSkip)

	c.JSON(200, usersData)
}

// CreateUser godoc
// @Summary      Create a new user
// @Description  Creates a user and returns the details
// @Tags         users
// @Accept       json
// @Produce      json
// @Param        user  body      types.AddUserRequestBody  true  "User data"
// @Success      201   {object}  map[string]interface{}
// @Failure      400   {object}  map[string]string
// @Failure      500
// @Router       /users [post]
func addUserHandler(c *gin.Context) {
	var addUserBody types.AddUserRequestBody
	json.NewDecoder(c.Request.Body).Decode(&addUserBody)

	isRequestBodyValid := validateRequestBody(addUserBody, c)

	if !isRequestBodyValid {
		return
	}

	newUserId := uuid.New().String()

	result, error := addUser(addUserBody, newUserId)

	if error != nil {
		c.Status(http.StatusInternalServerError)
		panic(error)
	}

	if !result.Acknowledged {
		c.Status(http.StatusInternalServerError)
		return
	}

	responseJson := gin.H{"Message": "New user created!", "id": newUserId}
	c.JSON(http.StatusCreated, responseJson)
}

// DeleteUser godoc
// @Summary      Delete a user
// @Description  Delete a user by their ID
// @Tags         users
// @Produce      json
// @Param        id   path      int  true  "User ID"
// @Success      204
// @Failure      404  {object}   map[string]string
// @Failure      500
// @Router       /users/{id} [delete]
func deleteUserHandler(c *gin.Context) {
	userId := extractUserId(c)
	result, error := deleteUser(userId)

	if error != nil {
		c.Status(http.StatusInternalServerError)
		panic(error)
	}

	if result.DeletedCount == 0 {
		c.JSON(http.StatusNotFound, gin.H{"Message": "No user with this id found !"})
		return
	}

	c.JSON(http.StatusNoContent, gin.H{"Message": "User deleted !"})
}

// UpdateUser godoc
// @Summary      Update a user
// @Description  Update a user's details by their ID
// @Tags         users
// @Accept       json
// @Produce      json
// @Param        id    path     int                       true  "User ID"
// @Param        user  body      types.UpdateUserRequestBody  true  "User data to update"
// @Success      200   {object}  map[string]string
// @Failure      404   {object}  map[string]string
// @Failure      500
// @Router       /users/{id} [patch]
func updateUserHandler(c *gin.Context) {
	var updateUserData types.UpdateUserRequestBody
	json.NewDecoder(c.Request.Body).Decode(&updateUserData)

	userId := extractUserId(c)
	result, error := updateUser(userId, updateUserData)

	if error != nil {
		c.Status(http.StatusInternalServerError)
		panic(error)
	}

	if result.MatchedCount == 0 {

		c.JSON(http.StatusNotFound, gin.H{"Message": "No user with this id found !"})
		return
	}

	c.JSON(http.StatusOK, gin.H{"Message": "User updated !"})

}